refactor(server): use Deno.serve instead of serve from std/http
The changes in this commit involve refactoring the server.ts file to use
Deno's built-in serve function instead of the serve function from the
standard http library. This change simplifies the code and reduces the
number of dependencies. Also, the way of getting the 'Content-Type' header
has been changed from 'c.req.headers.get' to 'c.req.header'.