feat(asciidoc): switch to asciidoctor-wasm for rendering
The AsciidocRenderer has been updated to use the asciidoctor-wasm
package instead of the @asciidoctor/core package. This change includes
the removal of the SourceMap function and the modification of the
Options type. The render method now uses a Ruby script to convert the
text. The create method has been updated to initialize the Asciidoctor
from a URL.